+++ examples/oberon2/src/oberon2_grammar_trait.rs
@@ -4,9 +4,9 @@
 // lost after next build.
 // ---------------------------------------------------------
 
-// Specifically disable clippy warnings that result in the way parol generates names.
-// The user can't influence this in its grammar definition.
+// Disable clippy warnings that can result in the way how parol generates code.
 #![allow(clippy::enum_variant_names)]
+#![allow(clippy::large_enum_variant)]
 #![allow(clippy::upper_case_acronyms)]
 
 use parol_runtime::derive_builder::Builder;